requirements.txt
loguru==0.7.2
openpyxl==3.1.2
markdown-it-py==3.0.0
pyyaml==6.0.1
typer==0.9.0
pyinstaller==6.0.0
